backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
motor==3.3.1
python-dotenv>=1.0.1
starlette>=0.36.0
python-multipart>=0.0.9
bcrypt==4.1.3
passlib>=1.7.4
python-jose>=3.3.0
cryptography>=42.0.8
tzdata>=2024.2
